+++ NEWS.md
@@ -1,9 +1,8 @@
 # tabde (dev version)
 
-* Add support for defining SQL constraints in table designs. Currently only
-  PRIMARY KEY is supported.
+* Add support for defining SQL constraints and periods in table designs.
 * Add support for csvy files (csv files with yaml header) that can contain SQL 
-  constraints.
+  constraints and periods.